+++ x/evm/types/params_test.go
@@ -18,7 +18,7 @@ func TestParamsValidate(t *testing.T) {
 		{"default", DefaultParams(), false},
 		{
 			"valid",
-			NewParams("ara", false, true, true, DefaultChainConfig(), extraEips),
+			NewParams("ara", false, true, true, DefaultChainConfig(), extraEips, nil),
 			false,
 		},
 		{
@@ -56,9 +56,13 @@ func TestParamsValidate(t *testing.T) {
 
 func TestParamsEIPs(t *testing.T) {
 	extraEips := []int64{2929, 1884, 1344}
-	params := NewParams("ara", false, true, true, DefaultChainConfig(), extraEips)
+	params := NewParams("ara", false, true, true, DefaultChainConfig(), extraEips, nil)
 	actual := params.EIPs()
 
+	require.NoError(t, validateChannels([]string{"channel-0"}))
+	require.Error(t, validateChannels(false))
+	require.Error(t, validateChannels(int64(123)))
+	require.Error(t, validateChannels(""))
 	require.Equal(t, []int{2929, 1884, 1344}, actual)
 }